requirements.txt
face-recognition==1.3.0
pillow==8.2.0
python-telegram-bot==12.2.0
